pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dma-traffic-gen"
version = "0.1.0"
description = "DMA traffic generator for Mobile SoC multimedia IPs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "click>=8.1",
  "pandas>=2.2",
  "plotly>=5.22",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "simpy>=4.1",
]

[project.optional-dependencies]
dev = [
  "hypothesis>=6.112",
  "pytest>=8.2",
]

[project.scripts]
dma-traffic-gen = "dma_traffic_gen.cli:main"

[tool.setuptools.packages.find]
include = ["dma_traffic_gen*"]
